Split image tagging screen into separate class
